merge :  정규 기능들은 개발을 마쳐서 main에 merge

+++ src/main/java/com/stool/studentcooperationtools/domain/room/service/RoomService.java
@@ -15,6 +15,7 @@ import com.stool.studentcooperationtools.domain.room.controller.request.RoomEnte
 import com.stool.studentcooperationtools.domain.room.controller.request.RoomRemoveRequest;
 import com.stool.studentcooperationtools.domain.room.controller.request.RoomTopicUpdateRequest;
 import com.stool.studentcooperationtools.domain.room.controller.response.RoomAddResponse;
+import com.stool.studentcooperationtools.domain.room.controller.response.RoomEnterResponse;
 import com.stool.studentcooperationtools.domain.room.controller.response.RoomSearchResponse;
 import com.stool.studentcooperationtools.domain.room.controller.response.RoomsFindResponse;
 import com.stool.studentcooperationtools.domain.room.repository.RoomRepository;
@@ -87,9 +88,10 @@ public class RoomService {
 
     @Transactional
     public Boolean removeRoom(SessionMember member, final RoomRemoveRequest request) {
-        Room room = roomRepository.findRoomByRoomId(member.getMemberSeq(), request.getRoomId())
-                .orElseThrow(() -> new IllegalArgumentException("소속되지 않은 방 정보입니다"));
+        Room room = roomRepository.findRoomWithPLock(request.getRoomId())
+                .orElseThrow(() -> new IllegalArgumentException("방 id 오류"));
         if(Objects.equals(member.getMemberSeq(), room.getLeader().getId())){
+            presentationService.deletePresentation(request.getRoomId());
             chatRepository.deleteByRoomId(room.getId());
             partRepository.deleteByRoomId(room.getId());
             topicRepository.deleteByRoomId(room.getId());
@@ -100,21 +102,36 @@ public class RoomService {
         else{
             Member teammate = memberRepository.findById(member.getMemberSeq())
                     .orElseThrow(() -> new IllegalArgumentException("유저 정보가 올바르지 않습니다"));
-            participationRepository.deleteByMemberIdAndRoomId(teammate.getId(), room.getId());
+            if(participationRepository.existsByMemberIdAndRoomId(teammate.getId(), room.getId())){
+                delParticipation(member, room);
+                participationRepository.deleteByMemberIdAndRoomId(teammate.getId(), room.getId());
+            }
+            else{
+                throw new IllegalArgumentException("참여 정보가 없는 유저입니다");
+            }
+
         }
-        presentationService.deletePresentation(request.getRoomId());
         return true;
     }
 
+    private void delParticipation(SessionMember member, Room room){
+        Member user = memberRepository.findById(member.getMemberSeq())
+                .orElseThrow(() -> new IllegalArgumentException("유저 정보가 올바르지 않습니다"));
+        Participation participation = participationRepository.findByMemberIdAndRoomId(user.getId(), room.getId());
+        room.deleteParticipation(participation);
+    }
+
     @Transactional
-    public boolean enterRoom(SessionMember member, final RoomEnterRequest request){
+    public RoomEnterResponse enterRoom(SessionMember member, final RoomEnterRequest request){
         Room room = roomRepository.findRoomWithPLock(request.getRoomId())
                 .orElseThrow(()-> new IllegalArgumentException("방 id 오류"));
         if(!room.verifyPassword(request.getPassword())) {
             throw new IllegalArgumentException("올바르지 않은 비밀번호입니다");
         }
         addParticipation(member,room);
-        return true;
+        return RoomEnterResponse.builder()
+                .leaderId(room.getLeader().getId())
+                .build();
     }
 
     private void addParticipation(SessionMember member, Room room){
